Initial commit
Setup the project workspace with library and header files. The provided
Makefile will compile and build the project **BUT** it must be ran from
a VC-compatible shell. This is because we rely on the Microsoft link.exe
to actually link the executable as LLVM does not provide a linker. The
file vcshell.cmd may be used to launch a VC-Compatible shell prompt.

Note: The GLFW equivalent of a HelloWorld program **REQUIRES** including
the GLFW header files, as the program will fail to link without a
specific definition of main.