Begin implementing voxel hashing
This is the first step towards a better voxel leaf data lookup scheme.
Currently, we use a giant sparse texture that serves as a sort of
hashmap from voxel leaf positions to attribute data. The original idea
was to emulate a hashmap on the GPU with the minimum effort required.
Unfortunately, most of the sparse texture cells are *really* sparse,
sitting at under 10% occupancy in many cases.

The new version should considerably improve this state of affairs by
storing voxel leaf data in a real, GPU-side hashmap. It is bassed on the
paper 'Real Time Hashing on the GPU' by Alcantera et al.